added knex folder, knex.js file, refactored seeds,
