+++ database/knex/knex.js
@@ -1,0 +1,3 @@
+const environment = process.env.ENVIRONMENT || "development";
+const config = require('../knexfile')[environment];
+module.exports = require("knex")(config);
