+++ server/resolvers.js
@@ -1,5 +1,5 @@
 const db = require("./db");
-const knex = require('../database/migrations/knexfile');
+const knex = require('../database/knexfile');
 
 
 const Query = {
